Add files via upload
The exercises for chapter 8 were not completed due to the heavy computational loads required by such algorithms used in this chapter. I was not motivated to set up parallel computing for this chapter as I have had previous experience using tree models, such as in the well-known tutorial presented by David Langer, analysing the Titanic data material on Kaggle.